fix: change balance data type to float64

+++ models/payment.go
@@ -2,15 +2,15 @@ package models
 
 // GetTransactionMutationRequest func
 type GetTransactionMutationRequest struct {
-	ID            string `json:"id"`
-	BankID        string `json:"bank_id"`
-	AccountNumber string `json:"account_number"`
-	BankType      string `json:"bank_type"`
-	Date          string `json:"date"`
-	Amount        int    `json:"amount"`
-	Description   string `json:"description"`
-	Type          string `json:"type"`
-	Balance       int    `json:"balance"`
+	ID            string  `json:"id"`
+	BankID        string  `json:"bank_id"`
+	AccountNumber string  `json:"account_number"`
+	BankType      string  `json:"bank_type"`
+	Date          string  `json:"date"`
+	Amount        int     `json:"amount"`
+	Description   string  `json:"description"`
+	Type          string  `json:"type"`
+	Balance       float64 `json:"balance"`
 }
 
 type DepositParameter struct {
